fix: HACS remove guards store-only repos, top-level success envelopes, module split
hacs/repository/remove reports success for a repository that was never
downloaded (uninstall no-ops without local files) — the tool now probes
installed state first and raises RESOURCE_NOT_FOUND. All five HACS
handlers hoist success above the add_timezone_metadata data wrapper per
the response-envelope contract, with the e2e extractor carrying the
top-level flag through. tools_hacs.py crossed the ~1000-line split
threshold; the scoring and registration-wait machinery moves to
hacs_registration.py with re-imports keeping established patch targets.
E2E: remove-nonexistent and store-only contracts plus a skip-guarded
download-remove roundtrip.
